# openapi.json
{"openapi":"3.1.0","info":{"title":"Supply Chain AI","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/upload":{"post":{"summary":"Upload File","description":"Accept an Excel file; return a session_id and dataset metadata.","operationId":"upload_file_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetMeta"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"summary":"Chat","description":"Run the agent + LLM and return the answer.","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/session/{session_id}":{"delete":{"summary":"Delete Session","description":"Clean up a session.","operationId":"delete_session_session__session_id__delete","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_file_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_upload_post"},"ChatRequest":{"properties":{"session_id":{"type":"string","title":"Session Id"},"question":{"type":"string","title":"Question"}},"type":"object","required":["session_id","question"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"raw_output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Output"}},"type":"object","required":["answer"],"title":"ChatResponse"},"DatasetMeta":{"properties":{"session_id":{"type":"string","title":"Session Id"},"filename":{"type":"string","title":"Filename"},"rows":{"type":"integer","title":"Rows"},"columns":{"type":"integer","title":"Columns"},"unique_suppliers":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Unique Suppliers"},"unique_materials":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Unique Materials"},"preview":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Preview"}},"type":"object","required":["session_id","filename","rows","columns","preview"],"title":"DatasetMeta"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}